+++ packages/fcl/src/serialize/index.js
@@ -3,12 +3,11 @@ import {resolve as defaultResolve} from "@onflow/sdk"
 import {config, createSignableVoucher} from "@onflow/sdk"
 
 export const serialize = async (args = [], opts = {}) => {
-  // prettier-ignore
-  const resolveFunction = await config().get(
-    "sdk.resolve",
+  const resolveFunction = await config.first(
+    ["sdk.resolve"],
     opts.resolve || defaultResolve
   )
-  
+
   if (Array.isArray(args)) args = await pipe(interaction(), args)
 
   return JSON.stringify(